Move committee resources
